extract mockCaptureClient as separate module, move tests
